Integrated SampleView and Shape handling
